Backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
email-validator
fastapi-clerk-auth
pytest
python-jose
PyJWT
pytest-asyncio
motor
pymongo
mongomock
httpx 
trio
pytest
requests
google-genai
tzdata
pytest-cov
selenium 
webdriver-manager